Always return an ArrayBuffer on success in SmartCardConnection.control
Some control codes do not return any data (hence the ArrayBuffer being
nullable). But You can represent this case with an empty ArrayBuffer
just as well and is more convenient then checking for null. Also, what
would be the semantic difference between null and an empty ArrayBuffer?
It's just ambiguous.