transaction function returns card disposition
This parameter is needed for the EndTransaction() PC/SC call.

It's a common practice to reset/unpower a card after exclusive use so
that other PC/SC applications cannot exploit the state a card was left
on.  An example is if the card was unlocked by a PIN entry during that
transaction.