Add event count information to GetStatusChange
This is an undocumented API but both Windows and PCSClite
implement it. PCSClite does document it, luckily. It's also not part of
the PC/SC spec. It's not present on MacOS though.

Moretheless, one does need to provide this information on Windows
in order for that function to work correctly, otherwise it will always
return immediately as the number of expected events (zero) won't match
the actual event count for that reader.

This information is "smuggled" in the upper two bytes of the
dwEventState and dwCurrentState member variables of the
SCARD_READERSTATE struct.