Add a sperate timeout parameter to getStatusChange()
The timeout parameter of SCardGetStatusChange() and SCardCancel() usage
are not equivalent in terms of resulting behavior.  This creates room
for behavioural differences and bugs when moving PC/SC C code to be run
on top of the Web API.

Here, for instance, a timeout of zero is passed:
https://github.com/OpenSC/OpenSC/blob/master/src/libopensc/reader-pcsc.c#L375
This cannot be implemented with SCardCancel.

When SCardCancel is called very close to a SCardGetStatusChange, there's
no guarantee it will work and it depends on the PC/SC implementation.
Even if the SCardGetStatusChange is processed before the SCardCancel,
the service might still be setting up some internal state for
SCardGetStatusChange and therefore that cancel might not work at that
early moment.

Thus a getStatusChange() with a very short AbortSignal.timeout(ms) (eg:
a couple of milliseconds) would most likely not get cancelled at all.