Expose SCARD_E_INVALID_PARAMETER as a TypeError
This is not only triggered by bad PC/SC API usage (eg, passing a null
context), which would be a browser implementation bug. It can also be
returned by the reader driver or card on input that, from a pure PC/SC
API perspective, is perfectly valid.

One example is when the user tries to transmit an APDU that is too big
for the smart card at hand (eg, card has not enough memory to hold it).